+++ client/dashboard/.eslintrc.js
@@ -32,6 +32,7 @@ module.exports = {
 							'@automattic/components/src/*',
 							'!@automattic/components/src/summary-button',
 							'!@automattic/components/src/core-badge',
+							'!@automattic/components/src/page-header',
 							'!@automattic/dataviews',
 							// Please do not add exceptions unless agreed on
 							// with the #architecture group.