+++ client/dashboard/agency-overview/index.tsx
@@ -1,22 +1,24 @@
+import { PageHeader } from '@automattic/components/src/page-header';
 import { Button } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import PageLayout from '../components/page-layout';
 
 export default function AgencyOverview() {
 	return (
-		<PageLayout
-			title={ __( 'Agency Overview' ) }
-			actions={
-				<>
-					<Button variant="primary" __next40pxDefaultSize>
+		<PageLayout>
+			<PageHeader
+				title={ __( 'Agency Overview' ) }
+				actions={ [
+					<Button key="add-sites" variant="primary" __next40pxDefaultSize>
 						{ __( 'Add Sites' ) }
-					</Button>
-					<Button variant="secondary" __next40pxDefaultSize>
+					</Button>,
+					<Button key="add-products" variant="secondary" __next40pxDefaultSize>
 						{ __( 'Add Products' ) }
-					</Button>
-				</>
-			}
-			description={ __( 'This is a sample overview page.' ) }
-		/>
+					</Button>,
+				] }
+				description={ __( 'This is a sample overview page.' ) }
+				level={ 1 }
+			/>
+		</PageLayout>
 	);
 }