+++ client/dashboard/app/404/index.tsx
@@ -1,18 +1,22 @@
+import { PageHeader } from '@automattic/components/src/page-header';
 import { __ } from '@wordpress/i18n';
 import PageLayout from '../../components/page-layout';
 import RouterLinkButton from '../../components/router-link-button';
 
 function NotFound() {
 	return (
-		<PageLayout
-			title={ __( '404 Not Found' ) }
-			description={ __( 'The page you are looking for does not exist.' ) }
-			actions={
-				<RouterLinkButton to="/sites" variant="primary" __next40pxDefaultSize>
-					{ __( 'Go to Sites' ) }
-				</RouterLinkButton>
-			}
-		/>
+		<PageLayout>
+			<PageHeader
+				title={ __( '404 Not Found' ) }
+				description={ __( 'The page you are looking for does not exist.' ) }
+				actions={ [
+					<RouterLinkButton key="go-to-sites" to="/sites" variant="primary" __next40pxDefaultSize>
+						{ __( 'Go to Sites' ) }
+					</RouterLinkButton>,
+				] }
+				level={ 1 }
+			/>
+		</PageLayout>
 	);
 }