+++ client/dashboard/app/500/index.tsx
@@ -1,3 +1,4 @@
+import { PageHeader } from '@automattic/components/src/page-header';
 import { Notice } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import PageLayout from '../../components/page-layout';
@@ -5,15 +6,17 @@ import RouterLinkButton from '../../components/router-link-button';
 
 function UnknownError( { error }: { error: Error } ) {
 	return (
-		<PageLayout
-			title={ __( '500 Error' ) }
-			description={ __( 'Something wrong happened.' ) }
-			actions={
-				<RouterLinkButton to="/sites" variant="primary" __next40pxDefaultSize>
-					{ __( 'Go to Sites' ) }
-				</RouterLinkButton>
-			}
-		>
+		<PageLayout>
+			<PageHeader
+				title={ __( '500 Error' ) }
+				level={ 1 }
+				description={ __( 'Something wrong happened.' ) }
+				actions={ [
+					<RouterLinkButton key="go-to-sites" to="/sites" variant="primary" __next40pxDefaultSize>
+						{ __( 'Go to Sites' ) }
+					</RouterLinkButton>,
+				] }
+			/>
 			<Notice status="error" isDismissible={ false }>
 				{ error.message }
 			</Notice>