+++ client/dashboard/domains/index.tsx
@@ -1,3 +1,4 @@
+import { PageHeader } from '@automattic/components/src/page-header';
 import { DataViews, filterSortAndPaginate, View } from '@automattic/dataviews';
 import { useQuery } from '@tanstack/react-query';
 import { Button } from '@wordpress/components';
@@ -99,14 +100,16 @@ function Domains() {
 	}
 	const { data: filteredData, paginationInfo } = filterSortAndPaginate( domains, view, fields );
 	return (
-		<PageLayout
-			title={ __( 'Domains' ) }
-			actions={
-				<Button variant="primary" __next40pxDefaultSize>
-					{ __( 'Add New Domain' ) }
-				</Button>
-			}
-		>
+		<PageLayout>
+			<PageHeader
+				title={ __( 'Domains' ) }
+				level={ 1 }
+				actions={ [
+					<Button key="add-new-domain" variant="primary" __next40pxDefaultSize>
+						{ __( 'Add New Domain' ) }
+					</Button>,
+				] }
+			/>
 			<DataViewsCard>
 				<DataViews
 					data={ filteredData || [] }