+++ client/dashboard/me/payment-methods/index.tsx
@@ -1,9 +1,11 @@
+import { PageHeader } from '@automattic/components/src/page-header';
 import { __ } from '@wordpress/i18n';
 import PageLayout from '../../components/page-layout';
 
 export default function PaymentMethods() {
 	return (
-		<PageLayout title={ __( 'Payment Methods' ) } size="small">
+		<PageLayout size="small">
+			<PageHeader title={ __( 'Payment Methods' ) } level={ 1 } />
 			<div>Payment methods content will go here</div>
 		</PageLayout>
 	);