+++ client/dashboard/me/privacy/index.tsx
@@ -1,13 +1,16 @@
+import { PageHeader } from '@automattic/components/src/page-header';
 import { __ } from '@wordpress/i18n';
 import PageLayout from '../../components/page-layout';
 
 function Privacy() {
 	return (
-		<PageLayout
-			title={ __( 'Privacy' ) }
-			description={ __( 'Manage your privacy settings.' ) }
-			size="small"
-		/>
+		<PageLayout size="small">
+			<PageHeader
+				title={ __( 'Privacy' ) }
+				description={ __( 'Manage your privacy settings.' ) }
+				level={ 1 }
+			/>
+		</PageLayout>
 	);
 }