+++ client/dashboard/sites/performance/index.tsx
@@ -1,8 +1,13 @@
+import { PageHeader } from '@automattic/components/src/page-header';
 import { __ } from '@wordpress/i18n';
 import PageLayout from '../../components/page-layout';
 
 function SitePerformance() {
-	return <PageLayout title={ __( 'Performance' ) } />;
+	return (
+		<PageLayout>
+			<PageHeader title={ __( 'Performance' ) } level={ 1 } />
+		</PageLayout>
+	);
 }
 
 export default SitePerformance;