+++ packages/components/CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - Add `FlowQuestion` component
 - Add `Breadcrumbs` and `Menu` components ([#102998](https://github.com/Automattic/wp-calypso/pull/102998)).
+- Add `PageHeader` component ([#103159](https://github.com/Automattic/wp-calypso/pull/103159)).
 - Add `SummaryButton` component ([#102827](https://github.com/Automattic/wp-calypso/pull/102827)).
 - Add `CoreBadge` component ([#103069](https://github.com/Automattic/wp-calypso/pull/103069)).
 - Add `Tabs` component ([#103231](https://github.com/Automattic/wp-calypso/pull/103231)).